ui: draw_fb: move the demo_mode back into the main file
This cleans up the source tree a bit.
It was also possible to shrink down the shim a bit by using existing
structs from the framebuffer crate.
